clean some things up and move the Packet class into the Network folder
