+++ FileteleporterTransfert/Server/Client.cs
@@ -7,6 +7,7 @@ using System.Numerics;
 using System.IO;
 using FileteleporterTransfert.Tools;
 using System.Threading.Tasks;
+using FileteleporterTransfert.Network;
 
 namespace server
 {
